[change] controllers location to versionnized namespace api/v1
